requirements.txt
google-adk
requests
python-dotenv
pypdf